requirements.txt
streamlit
openai
llama-index>=0.9.48
python-dotenv
